requirements.txt
absl-py==2.1.0
altair==5.5.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cmake==4.0.0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
facenet-pytorch==2.6.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.32.3
idna==3.10
imageio==2.37.0
imbalanced-learn==0.13.0
jax==0.5.0
jaxlib==0.5.0
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.0
mediapipe==0.10.21
ml_dtypes==0.5.1
mpmath==1.3.0
narwhals==1.41.0
networkx==3.5
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
pillow==10.2.0
protobuf==4.25.6
pyarrow==20.0.0
pycparser==2.22
pydeck==0.9.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.25.1
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
sentencepiece==0.2.0
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
sounddevice==0.5.1
streamlit==1.45.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.6.1
tokenizers==0.21.1
toml==0.10.2
torch==2.2.2
torchvision==0.17.2
tornado==6.5.1
tqdm==4.67.1
transformers==4.52.4
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
xgboost==3.0.0
